Remove depreciated functions
