+++ eqcorrscan/core/template_gen.py
@@ -558,6 +558,19 @@ def _download_from_client(client, client_type, catalog, data_pad, process_len,
     return st
 
 
+def _rms(array):
+    """
+    Calculate RMS of array.
+
+    :type array: numpy.ndarray
+    :param array: Array to calculate the RMS for.
+
+    :returns: RMS of array
+    :rtype: float
+    """
+    return np.sqrt(np.mean(np.square(array)))
+
+
 def _template_gen(picks, st, length, swin='all', prepick=0.05,
                   all_horiz=False, delayed=True, plot=False, min_snr=None):
     """
@@ -635,7 +648,6 @@ def _template_gen(picks, st, length, swin='all', prepick=0.05,
     """
     from eqcorrscan.utils.plotting import pretty_template_plot as tplot
     from eqcorrscan.utils.plotting import noise_plot
-    from eqcorrscan.core.bright_lights import _rms
 
     # the users picks intact.
     if not isinstance(swin, list):
@@ -834,446 +846,6 @@ def _group_events(catalog, process_len, template_length, data_pad):
             sub_catalog = Catalog([event])
     sub_catalogs.append(sub_catalog)
     return sub_catalogs
-
-
-# TODO: Remove these depreciated functions
-def multi_template_gen(catalog, st, length, swin='all', prepick=0.05,
-                       all_horiz=False, delayed=True, plot=False,
-                       return_event=False, min_snr=None):
-    """
-    Generate multiple templates from one stream of data.
-
-    Thin wrapper around _template_gen to generate multiple templates from
-    one stream of continuous data.  Takes processed (filtered and resampled)
-    seismic data!
-
-    :type catalog: obspy.core.event.Catalog
-    :param catalog: Events to extract templates for
-    :type st: obspy.core.stream.Stream
-    :param st:
-        Processed stream to extract from, e.g. filtered and re-sampled to what
-        you want using pre_processing.dayproc.
-    :type length: float
-    :param length: Length of template in seconds
-    :type swin: string
-    :param swin:
-        P, S, P_all, S_all or all, defaults to all: see note in
-        :func:`eqcorrscan.core.template_gen.template_gen`
-    :type prepick: float
-    :param prepick:
-        Length in seconds to extract before the pick time default is
-        0.05 seconds.
-    :type all_horiz: bool
-    :param all_horiz:
-        To use both horizontal channels even if there is only a pick on one of
-        them.  Defaults to False.
-    :type delayed: bool
-    :param delayed:
-        If True, each channel will begin relative to it's own pick-time, if set
-         to False, each channel will begin at the same time.
-    :type plot: bool
-    :param plot: To plot the template or not, default is True
-    :type return_event: bool
-    :param return_event: Whether to return the event and process length or not.
-    :type min_snr: float
-    :param min_snr:
-        Minimum signal-to-noise ratio for a channel to be included in the
-        template, where signal-to-noise ratio is calculated as the ratio of
-        the maximum amplitude in the template window to the rms amplitude in
-        the whole window given.
-
-    :returns: List of :class:`obspy.core.stream.Stream` templates.
-    :rtype: list
-
-    .. warning::
-        Data must be processed before using this function - highcut, lowcut and
-        filt_order are only used to generate the meta-data for the templates.
-
-    .. note:: By convention templates are generated with P-phases on the \
-        vertical channel and S-phases on the horizontal channels, normal \
-        seismograph naming conventions are assumed, where Z denotes vertical \
-        and N, E, R, T, 1 and 2 denote horizontal channels, either oriented \
-        or not.  To this end we will **only** use Z channels if they have a \
-        P-pick, and will use one or other horizontal channels **only** if \
-        there is an S-pick on it.
-
-    .. warning:: If there is no phase_hint included in picks, and swin=all, \
-        all channels with picks will be used.
-    """
-    EQcorrscanDeprecationWarning(
-        "Function is depreciated and will be removed soon. Use "
-        "template_gen.template_gen instead.")
-    temp_list = template_gen(
-        method="from_meta_file", process=False, meta_file=catalog, st=st,
-        lowcut=None, highcut=None, samp_rate=st[0].stats.sampling_rate,
-        filt_order=None, length=length, prepick=prepick,
-        swin=swin, all_horiz=all_horiz, delayed=delayed, plot=plot,
-        return_event=return_event, min_snr=min_snr, parallel=False)
-    return temp_list
-
-
-def from_client(catalog, client_id, lowcut, highcut, samp_rate, filt_order,
-                length, prepick, swin, process_len=86400, data_pad=90,
-                all_horiz=False, delayed=True, plot=False, return_event=False,
-                min_snr=None):
-    """
-    Generate multiplexed template from FDSN client.
-
-    Function to generate templates from an FDSN client. Must be given \
-    an obspy.Catalog class and the client_id as input. The function returns \
-    a list of obspy.Stream classes containing steams for each desired \
-    template.
-
-    :type catalog: obspy.core.event.Catalog
-    :param catalog: Catalog class containing desired template events
-    :type client_id: str
-    :param client_id: Name of the client, either url, or Obspy \
-        mappable (see the :mod:`obspy.clients.fdsn` documentation).
-    :type lowcut: float
-    :param lowcut: Low cut (Hz), if set to None will not apply a lowcut.
-    :type highcut: float
-    :param highcut: High cut (Hz), if set to None will not apply a highcut.
-    :type samp_rate: float
-    :param samp_rate: New sampling rate in Hz.
-    :type filt_order: int
-    :param filt_order: Filter level (number of corners).
-    :type length: float
-    :param length: Extract length in seconds.
-    :type prepick: float
-    :param prepick: Pre-pick time in seconds
-    :type swin: str
-    :param swin:
-        P, S, P_all, S_all or all, defaults to all: see note in
-        :func:`eqcorrscan.core.template_gen.template_gen`
-    :type process_len: int
-    :param process_len: Length of data in seconds to download and process.
-    :param data_pad: Length of data (in seconds) required before and after \
-        any event for processing, use to reduce edge-effects of filtering on \
-        the templates.
-    :type data_pad: int
-    :type all_horiz: bool
-    :param all_horiz: To use both horizontal channels even if there is only \
-        a pick on one of them.  Defaults to False.
-    :type delayed: bool
-    :param delayed: If True, each channel will begin relative to it's own \
-        pick-time, if set to False, each channel will begin at the same time.
-    :type plot: bool
-    :param plot: Plot templates or not.
-    :type return_event: bool
-    :param return_event: Whether to return the event and process length or not.
-    :type min_snr: float
-    :param min_snr:
-        Minimum signal-to-noise ratio for a channel to be included in the
-        template, where signal-to-noise ratio is calculated as the ratio of
-        the maximum amplitude in the template window to the rms amplitude in
-        the whole window given.
-
-    :returns: List of :class:`obspy.core.stream.Stream` Templates
-    :rtype: list
-
-    .. warning::
-        This function is depreciated and will be removed in a forthcoming
-        release. Please use `template_gen` instead.
-
-    .. note::
-        process_len should be set to the same length as used when computing
-        detections using match_filter.match_filter, e.g. if you read
-        in day-long data for match_filter, process_len should be 86400.
-
-    .. rubric:: Example
-
-    >>> from obspy.clients.fdsn import Client
-    >>> from eqcorrscan.core.template_gen import from_client
-    >>> client = Client('NCEDC')
-    >>> catalog = client.get_events(eventid='72572665', includearrivals=True)
-    >>> # We are only taking two picks for this example to speed up the
-    >>> # example, note that you don't have to!
-    >>> catalog[0].picks = catalog[0].picks[0:2]
-    >>> templates = from_client(catalog=catalog, client_id='NCEDC',
-    ...                         lowcut=2.0, highcut=9.0, samp_rate=20.0,
-    ...                         filt_order=4, length=3.0, prepick=0.15,
-    ...                         swin='all', process_len=300,
-    ...                         all_horiz=True)
-    >>> templates[0].plot(equal_scale=False, size=(800,600)) # doctest: +SKIP
-
-    .. figure:: ../../plots/template_gen.from_client.png
-    """
-    EQcorrscanDeprecationWarning(
-        "Function is depreciated and will be removed soon. Use "
-        "template_gen.template_gen instead.")
-    temp_list = template_gen(
-        method="from_client", catalog=catalog, client_id=client_id,
-        lowcut=lowcut, highcut=highcut, samp_rate=samp_rate,
-        filt_order=filt_order, length=length, prepick=prepick,
-        swin=swin, process_len=process_len, data_pad=data_pad,
-        all_horiz=all_horiz, delayed=delayed, plot=plot,
-        return_event=return_event, min_snr=min_snr)
-    return temp_list
-
-
-def from_seishub(catalog, url, lowcut, highcut, samp_rate, filt_order,
-                 length, prepick, swin, process_len=86400, data_pad=90,
-                 all_horiz=False, delayed=True, plot=False,
-                 return_event=False, min_snr=None):
-    """
-    Generate multiplexed template from SeisHub database.
-
-    Function to generate templates from a SeisHub database. Must be given
-    an obspy.Catalog class and the SeisHub url as input. The function returns
-    a list of obspy.Stream classes containting steams for each desired
-    template.
-
-    :type catalog: :class:`obspy.core.event.Catalog`
-    :param catalog: Catalog class containing desired template events
-    :type url: str
-    :param url: url of SeisHub database instance
-    :type lowcut: float
-    :param lowcut: Low cut (Hz), if set to None will not apply a lowcut.
-    :type highcut: float
-    :param highcut: High cut (Hz), if set to None will not apply a highcut.
-    :type samp_rate: float
-    :param samp_rate: New sampling rate in Hz.
-    :type filt_order: int
-    :param filt_order: Filter level (number of corners).
-    :type length: float
-    :param length: Extract length in seconds.
-    :type prepick: float
-    :param prepick: Pre-pick time in seconds
-    :type swin: str
-    :param swin:
-        P, S, P_all, S_all or all, defaults to all: see note in
-        :func:`eqcorrscan.core.template_gen.template_gen`
-    :type process_len: int
-    :param process_len: Length of data in seconds to download and process.
-    :param data_pad:
-        Length of data (in seconds) required before and after any event for
-        processing, use to reduce edge-effects of filtering on the templates.
-    :type data_pad: int
-    :type all_horiz: bool
-    :param all_horiz:
-        To use both horizontal channels even if there is only a pick on one
-        of them.  Defaults to False.
-    :type delayed: bool
-    :param delayed:
-        If True, each channel will begin relative to it's own pick-time, if
-        set to False, each channel will begin at the same time.
-    :type plot: bool
-    :param plot: Plot templates or not.
-    :type return_event: bool
-    :param return_event: Whether to return the event and process length or not.
-    :type min_snr: float
-    :param min_snr:
-        Minimum signal-to-noise ratio for a channel to be included in the
-        template, where signal-to-noise ratio is calculated as the ratio of
-        the maximum amplitude in the template window to the rms amplitude in
-        the whole window given.
-
-    :returns: List of :class:`obspy.core.stream.Stream` of newly cut templates
-    :rtype: list
-
-    .. note::
-        process_len should be set to the same length as used when computing
-        detections using match_filter.match_filter, e.g. if you read
-        in day-long data fro match_filter, process_len should be 86400.
-
-    .. warning::
-        Not tested in continuous integration (due to lack of seishub client),
-        let us know of any failures.
-    """
-    EQcorrscanDeprecationWarning(
-        "Function is depreciated and will be removed soon. Use "
-        "template_gen.template_gen instead.")
-    temp_list = template_gen(
-        method="from_seishub", catalog=catalog, url=url,
-        lowcut=lowcut, highcut=highcut, samp_rate=samp_rate,
-        filt_order=filt_order, length=length, prepick=prepick,
-        swin=swin, process_len=process_len, data_pad=data_pad,
-        all_horiz=all_horiz, delayed=delayed, plot=plot,
-        return_event=return_event, min_snr=min_snr)
-    return temp_list
-
-
-def from_sac(sac_files, lowcut, highcut, samp_rate, filt_order, length, swin,
-             prepick, all_horiz=False, delayed=True, plot=False,
-             return_event=False, min_snr=None):
-    """
-    Generate a multiplexed template from a list of SAC files.
-
-    Function to read picks and waveforms from SAC data, and generate a \
-    template from these. Usually sac_files is a list of all single-channel \
-    SAC files for a given event, a single, multi-channel template will be \
-    created from these traces.
-
-    **All files listed in sac_files should be associated with a single event.**
-
-    :type sac_files: list
-    :param sac_files: osbpy.core.stream.Stream of sac waveforms, or
-        list of paths to sac waveforms.
-    :type lowcut: float
-    :param lowcut: Low cut (Hz), if set to None will not apply a lowcut.
-    :type highcut: float
-    :param highcut: High cut (Hz), if set to None will not apply a highcut.
-    :type samp_rate: float
-    :param samp_rate: New sampling rate in Hz.
-    :type filt_order: int
-    :param filt_order: Filter level.
-    :type length: float
-    :param length: Extract length in seconds.
-    :type swin: str
-    :param swin:
-        P, S, P_all, S_all or all, defaults to all: see note in
-        :func:`eqcorrscan.core.template_gen.template_gen`
-    :type prepick: float
-    :param prepick: Length to extract prior to the pick in seconds.
-    :type all_horiz: bool
-    :param all_horiz: To use both horizontal channels even if there is only \
-        a pick on one of them.  Defaults to False.
-    :type delayed: bool
-    :param delayed: If True, each channel will begin relative to it's own \
-        pick-time, if set to False, each channel will begin at the same time.
-    :type plot: bool
-    :param plot: Turns template plotting on or off.
-    :type return_event: bool
-    :param return_event: Whether to return the event and process length or not.
-    :type min_snr: float
-    :param min_snr:
-        Minimum signal-to-noise ratio for a channel to be included in the
-        template, where signal-to-noise ratio is calculated as the ratio of
-        the maximum amplitude in the template window to the rms amplitude in
-        the whole window given.
-
-    :returns: Newly cut template.
-    :rtype: :class:`obspy.core.stream.Stream`
-
-    .. note:: This functionality is not supported for obspy versions below \
-        1.0.0 as references times are not read in by SACIO, which are needed \
-        for defining pick times.
-
-    .. rubric:: Example
-
-    >>> from eqcorrscan.core.template_gen import from_sac
-    >>> import glob
-    >>> # Get the path to the test data
-    >>> import eqcorrscan
-    >>> import os
-    >>> TEST_PATH = os.path.dirname(eqcorrscan.__file__) + '/tests/test_data'
-    >>> # Get all the SAC-files associated with one event.
-    >>> sac_files = glob.glob(TEST_PATH + '/SAC/2014p611252/*')
-    >>> templates = from_sac(sac_files=sac_files, lowcut=2.0, highcut=10.0,
-    ...                      samp_rate=25.0, filt_order=4, length=2.0,
-    ...                      swin='all', prepick=0.1, all_horiz=True)
-    >>> print(templates[0][0].stats.sampling_rate)
-    25.0
-    >>> print(len(templates[0]))
-    15
-    """
-    EQcorrscanDeprecationWarning(
-        "Function is depreciated and will be removed soon. Use "
-        "template_gen.template_gen instead.")
-    temp_list = template_gen(
-        method="from_sac", sac_files=sac_files,
-        lowcut=lowcut, highcut=highcut, samp_rate=samp_rate,
-        filt_order=filt_order, length=length, prepick=prepick,
-        swin=swin, all_horiz=all_horiz, delayed=delayed, plot=plot,
-        return_event=return_event, min_snr=min_snr, parallel=False)
-    return temp_list
-
-
-def from_meta_file(meta_file, st, lowcut, highcut, samp_rate, filt_order,
-                   length, prepick, swin, all_horiz=False, delayed=True,
-                   plot=False, parallel=True, return_event=False,
-                   min_snr=None):
-    """
-    Generate a multiplexed template from a local file.
-
-    Function to generate a template from a local observation file
-    and an obspy.Stream object.
-
-    :type meta_file: str
-    :param meta_file: File containing pick information, can contain \
-        multiple events.  File must be formatted in a way readable by \
-        :func:`obspy.core.event.read_events`.
-    :type st: obspy.core.stream.Stream
-    :param st: Stream containing waveform data for template (hopefully). \
-        Note that this should be the same length of stream as you will use \
-        for the continuous detection, e.g. if you detect in day-long files, \
-        give this a day-long file!
-    :type lowcut: float
-    :param lowcut: Low cut (Hz), if set to None will not apply a lowcut.
-    :type highcut: float
-    :param highcut: High cut (Hz), if set to None will not apply a highcut.
-    :type samp_rate: float
-    :param samp_rate: New sampling rate in Hz.
-    :type filt_order: int
-    :param filt_order: Filter level (number of corners).
-    :type length: float
-    :param length: Extract length in seconds.
-    :type prepick: float
-    :param prepick: Pre-pick time in seconds
-    :type swin: str
-    :param swin:
-        P, S, P_all, S_all or all, defaults to all: see note in
-        :func:`eqcorrscan.core.template_gen.template_gen`
-    :type all_horiz: bool
-    :param all_horiz: To use both horizontal channels even if there is only \
-        a pick on one of them.  Defaults to False.
-    :type delayed: bool
-    :param delayed: If True, each channel will begin relative to it's own \
-        pick-time, if set to False, each channel will begin at the same time.
-    :type plot: bool
-    :param plot: Display template plots or not
-    :type parallel: bool
-    :param parallel: Whether to process data in parallel or not.
-    :type return_event: bool
-    :param return_event: Whether to return the event and process length or not.
-    :type min_snr: float
-    :param min_snr:
-        Minimum signal-to-noise ratio for a channel to be included in the
-        template, where signal-to-noise ratio is calculated as the ratio of
-        the maximum amplitude in the template window to the rms amplitude in
-        the whole window given.
-
-    :returns: List of :class:`obspy.core.stream.Stream` newly cut templates
-    :rtype: list
-
-    .. Note::
-        All picks must be associated with a station and channel, this is
-        not the case for NonLinLoc HYP files, will not use any picks that
-        do not have this association.
-
-    .. warning:: We suggest giving this function a full day of data, to \
-        ensure templates are generated with **exactly** the same processing \
-        as the continuous data.  Not doing this will result in slightly \
-        reduced cross-correlation values.
-
-    .. rubric:: Example
-
-    >>> from obspy import read
-    >>> from eqcorrscan.core.template_gen import from_meta_file
-    >>> # Get the path to the test data
-    >>> import eqcorrscan
-    >>> import os
-    >>> TEST_PATH = os.path.dirname(eqcorrscan.__file__) + '/tests/test_data'
-    >>> st = read(TEST_PATH + '/WAV/TEST_/' +
-    ...           '2013-09-01-0410-35.DFDPC_024_00')
-    >>> quakeml = TEST_PATH + '/20130901T041115.xml'
-    >>> templates = from_meta_file(meta_file=quakeml, st=st, lowcut=2.0,
-    ...                            highcut=9.0, samp_rate=20.0, filt_order=3,
-    ...                            length=2, prepick=0.1, swin='S',
-    ...                            all_horiz=True)
-    >>> print(len(templates[0]))
-    10
-    """
-    EQcorrscanDeprecationWarning(
-        "Function is depreciated and will be removed soon. Use "
-        "template_gen.template_gen instead.")
-    temp_list = template_gen(
-        method="from_meta_file", meta_file=meta_file, st=st,
-        lowcut=lowcut, highcut=highcut, samp_rate=samp_rate,
-        filt_order=filt_order, length=length, prepick=prepick,
-        swin=swin, all_horiz=all_horiz, delayed=delayed, plot=plot,
-        return_event=return_event, min_snr=min_snr, parallel=parallel)
-    return temp_list
 
 
 if __name__ == "__main__":
