--- eqcorrscan/doc/submodules/core.bright_lights.rst
@@ -1,38 +1,0 @@
-bright_lights
--------------
-
-.. currentmodule:: eqcorrscan.core.bright_lights
-.. automodule:: eqcorrscan.core.bright_lights
-
-    .. comment to end block
-
-    Classes & Functions
-    -------------------
-    .. autosummary::
-       :toctree: autogen
-       :nosignatures:
-
-       brightness
-       coherence
-
-    .. comment to end block
-
-    Private Functions
-    -----------------
-    Note that these functions are not designed for public use and may change
-    at any point.
-
-    .. autosummary::
-      :toctree: autogen
-      :nosignatures:
-
-      _rms
-      _rm_similarlags
-      _resample_grid
-      _read_tt
-      _cum_net_resp
-      _node_loop
-      _find_detections
-
-
-    .. comment to end block
