--- eqcorrscan/doc/submodules/utils.seismo_logs.rst
@@ -1,20 +1,0 @@
-seismo_logs
------------
-
-.. currentmodule:: eqcorrscan.utils.seismo_logs
-.. automodule:: eqcorrscan.utils.seismo_logs
-
-    .. comment to end block
-
-    Classes & Functions
-    -------------------
-    .. autosummary::
-       :toctree: autogen
-       :nosignatures:
-
-       check_all_logs
-       flag_time_err
-       rt_location_log
-       rt_time_log
-
-    .. comment to end block
