+++ eqcorrscan/doc/utils.rst
@@ -19,7 +19,6 @@ and for the analysis of waveforms detected by cross-correlation.
    submodules/utils.plotting
    submodules/utils.pre_processing
    submodules/utils.sac_util
-   submodules/utils.seismo_logs
    submodules/utils.stacking
    submodules/utils.synth_seis
    submodules/utils.trigger
