+++ eqcorrscan/utils/plotting.py
@@ -1543,101 +1543,6 @@ def plot_repicked(template, picks, det_stream, size=(10.5, 7.5), **kwargs):
     return fig
 
 
-def NR_plot(stream, NR_stream, detections, false_detections=False,
-            size=(18.5, 10), **kwargs):
-    """
-    Plot Network response alongside the stream used.
-
-    Highlights detection times in the network response.
-
-    :type stream: obspy.core.stream.Stream
-    :param stream: Stream to plot
-    :type NR_stream: obspy.core.stream.Stream
-    :param NR_stream: Stream for the network response
-    :type detections: list
-    :param detections: List of the detection time as :class:`datetime.datetime`
-    :type false_detections: list
-    :param false_detections:
-        Either False (default) or list of false detection times
-        (:class:`datetime.datetime`).
-    :type size: tuple
-    :param size: Size of figure, default is (18.5, 10)
-
-    :returns: :class:`matplotlib.figure.Figure`
-
-    .. Note::
-        Called by :mod:`eqcorrscan.core.bright_lights`, not a general use
-        plot (hence no example)
-    """
-    import matplotlib.pyplot as plt
-    fig, axes = plt.subplots(len(stream) + 1, 1, sharex=True, figsize=size)
-    if len(stream) > 1:
-        axes = axes.ravel()
-    else:
-        return
-    mintime = stream.sort(['starttime'])[0].stats.starttime
-    stream.sort(['network', 'station', 'starttime'])
-    for i, tr in enumerate(stream):
-        delay = tr.stats.starttime - mintime
-        delay *= tr.stats.sampling_rate
-        y = tr.data
-        x = [tr.stats.starttime.datetime + dt.timedelta(
-            seconds=s / tr.stats.sampling_rate) for s in range(len(y))]
-        x = mdates.date2num(x)
-        axes[i].plot(x, y, 'k', linewidth=1.1)
-        axes[i].set_ylabel('.'.join([tr.stats.station, tr.stats.channel]),
-                           rotation=0)
-        axes[i].yaxis.set_ticks([])
-        axes[i].set_xlim(x[0], x[-1])
-    # Plot the network response
-    tr = NR_stream[0]
-    delay = tr.stats.starttime - mintime
-    delay *= tr.stats.sampling_rate
-    y = tr.data
-    x = [tr.stats.starttime.datetime +
-         dt.timedelta(seconds=s / tr.stats.sampling_rate)
-         for s in range(len(y))]
-    x = mdates.date2num(x)
-    axes[-1].plot(x, y, 'k', linewidth=1.1)
-    axes[-1].set_ylabel('.'.join([tr.stats.station, tr.stats.channel]),
-                        rotation=0)
-    axes[-1].yaxis.set_ticks([])
-    axes[-1].set_xlabel('Time')
-    axes[-1].set_xlim(x[0], x[-1])
-    # Plot the detections!
-    ymin, ymax = axes[-1].get_ylim()
-    if false_detections:
-        for detection in false_detections:
-            xd = mdates.date2num(detection)
-            axes[-1].plot((xd, xd), (ymin, ymax), 'k--', linewidth=0.9,
-                          alpha=0.5)
-    for detection in detections:
-        xd = mdates.date2num(detection)
-        axes[-1].plot((xd, xd), (ymin, ymax), 'r--', linewidth=1.1)
-    # Set formatters for x-labels
-    mins = mdates.MinuteLocator()
-    timedif = tr.stats.endtime.datetime - tr.stats.starttime.datetime
-    if timedif.total_seconds() >= 10800 and timedif.total_seconds() <= 25200:
-        hours = mdates.MinuteLocator(byminute=[0, 15, 30, 45])
-    elif timedif.total_seconds() <= 1200:
-        hours = mdates.MinuteLocator(byminute=range(0, 60, 2))
-    elif timedif.total_seconds > 25200 and timedif.total_seconds() <= 172800:
-        hours = mdates.HourLocator(byhour=range(0, 24, 3))
-    elif timedif.total_seconds() > 172800:
-        hours = mdates.DayLocator()
-    else:
-        hours = mdates.MinuteLocator(byminute=range(0, 60, 5))
-    hrFMT = mdates.DateFormatter('%Y/%m/%d %H:%M:%S')
-    axes[-1].xaxis.set_major_locator(hours)
-    axes[-1].xaxis.set_major_formatter(hrFMT)
-    axes[-1].xaxis.set_minor_locator(mins)
-    plt.gcf().autofmt_xdate()
-    axes[-1].fmt_xdata = mdates.DateFormatter('%Y/%m/%d %H:%M:%S')
-    plt.subplots_adjust(hspace=0)
-    fig = _finalise_figure(fig=fig, **kwargs)  # pragma: no cover
-    return fig
-
-
 def svd_plot(svstreams, svalues, stachans, **kwargs):
     """
     Plot singular vectors from the :mod:`eqcorrscan.utils.clustering` routines.
